requirements.txt
asgiref==3.5.2
attrs==22.1.0
Authlib==1.0.1
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
